Separate connection pool timeout from request timeout and include separate config point for it
In cases where applications have large thread pools driving connections to Koji, but the
Koji connection pool is relatively small (to help avoid swamping Koji, for instance),
having the connection pool timeout set the same as the request timeout doesn't make sense.

In many cases, the requests in question may take a small fraction of the time it would
take to pare down the waiting-threads list, leading to either artificially high request
timeouts, or a lot of ConnectionTimeoutException errors.

This change fixes the problem by separating these configuration points, and setting
the default connection pool timeout to 60 seconds (the request timeout default is 30).